# Conflicts:
#	Makefile
#	compcert/Makefile
#	compcert/common/Events.v
#	concurrency/ClightCoreSemantincsForMachines.v
#	concurrency/HybridMachine_simulation.v
#	concurrency/call_lemmas.v
#	concurrency/disjointness.v
#	concurrency/inj_lemmas.v
#	concurrency/join_sm.v
#	concurrency/lifting.v
#	concurrency/lifting_safety.v
#	concurrency/linking.v
#	concurrency/linking_inv.v
#	concurrency/linking_proof.v
#	concurrency/linking_spec.v
#	concurrency/machine_simulation.v
#	concurrency/main.v
#	concurrency/reach_lemmas.v
#	concurrency/ret_lemmas.v
#	concurrency/semax_preservation_release.v
#	concurrency/sepcomp.v
#	concurrency/x86_context.v
#	concurrency/x86_inj.v
#	concurrency/x86_safe.v
#	sepcomp/simulations.v
#	veric/Clight_core.v
#	veric/Clight_new.v
#	veric/Clight_sim.v
#	veric/SeparationLogicSoundness.v
#	veric/SequentialClight.v
#	veric/semax_prog.v

+++ progs/reverse.c
@@ -1,15 +1,15 @@
 #include <stddef.h>
 
-struct list {int head; struct list *tail;};
+struct list {unsigned head; struct list *tail;};
 
 struct list three[] = {
     {1, three+1}, {2, three+2}, {3, NULL}
 };
 
-int sumlist (struct list *p) {
-  int s = 0;
+unsigned sumlist (struct list *p) {
+  unsigned s = 0;
   struct list *t = p;
-  int h;
+  unsigned h;
   while (t) {
      h = t->head;
      t = t->tail;
@@ -32,9 +32,9 @@ struct list *reverse (struct list *p) {
 }
 
 int main (void) {
-  struct list *r; int s;
+  struct list *r; unsigned s;
   r = reverse(three);
   s = sumlist(r);
-  return s;
+  return (int)s;
 }
